requirements.txt
flask_sqlalchemy
flask
flask_migrate
sqlalchemy
Flask-WTF
wtforms_sqlalchemy